requirements.txt
async-generator==1.10
attrs==22.2.0
blinker==1.4
Bootstrap-Flask==1.2.0
certifi==2022.12.7
click==7.1.1
coverage==7.2.2
distlib==0.3.6
exceptiongroup==1.1.1
Faker==4.0.2
filelock==3.9.0
Flask==1.1.2
Flask-Avatars==0.2.2
Flask-Dropzone==1.5.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.9.0
Flask-SQLAlchemy==1.0
flask-whooshee==0.7.0
Flask-WTF==0.14.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.2.0
kafka-python==2.0.2
MarkupSafe==1.1.1
mock==5.0.1
numpy==1.23
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.3.0
pipenv==2023.2.4
platformdirs==3.0.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.2
pytest-cov==4.0.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2022.7.1
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.0
selenium==4.8.2
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.7
surprise==0.1
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
urllib3==1.26.15
virtualenv==20.19.0
virtualenv-clone==0.5.7
watchdog==0.10.2
Werkzeug==2.0.3
Whoosh==2.7.4
wsproto==1.2.0
WTForms==2.2.1
zipp==3.13.0
psycopg2-binary==2.9.5
docker-compose==1.29.2
mlflow==2.2.2